updated dtype

+++ MERFISH_probe_design/probe_design/filters.py
@@ -52,7 +52,7 @@ def calc_tm_for_probe_dict(probe_dict:pd.core.frame.DataFrame, Na_conc:float, fm
                 
                 tms.append(MeltingTemp.chem_correction(tm_raw, fmd=fmd_percentile))
 
-            probe_dict[gk][tk][column_key_write] = pd.Series(tms, index=probe_dict[gk][tk].index)
+            probe_dict[gk][tk][column_key_write] = pd.Series(tms, index=probe_dict[gk][tk].index, dtype=np.float32)
 
 def calc_tm_JM(sequence:str, monovalentSalt:float=0.3, probeConc:float=5e-9):
     '''TM calculation used in JM's original MATLAB code.
